toolkit/devtools/server/tests/mochitest/test_inspector-insert.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript;version=1.8" src="inspector-helpers.js"></script>
  <script type="application/javascript;version=1.8">
Components.utils.import("resource://gre/modules/devtools/Loader.jsm");

const promise = devtools.require("sdk/core/promise");
const inspector = devtools.require("devtools/server/actors/inspector");

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
}

var gWalker = null;
var gClient = null;

function assertOwnership() {
  return assertOwnershipTrees(gWalker);
}

addTest(function setup() {
  let url = document.getElementById("inspectorContent").href;
  attachURL(url, function(err, client, tab, doc) {
    gInspectee = doc;
    let {InspectorFront} = devtools.require("devtools/server/actors/inspector");
    let inspector = InspectorFront(client, tab);
    promiseDone(inspector.getWalker().then(walker => {
      ok(walker, "getWalker() should return an actor.");
      gClient = client;
      gWalker = walker;
    }).then(runNextTest));
  });
});

addTest(function testRearrange() {
  let longlist = null;
  let nodeA = null;
  let nextNode = null;

  promiseDone(gWalker.querySelector(gWalker.rootNode, "#longlist").then(listFront => {
    longlist = listFront;
  }).then(() => {
    return gWalker.children(longlist);
  }).then(response => {
    nodeA = response.nodes[0];
    is(nodeA.id, "a", "Got the expected node.");
    // Move nodeA to the end of the list.
    return gWalker.insertBefore(nodeA, longlist, null);
  }).then(() => {
    ok(!gInspectee.querySelector("#a").nextSibling, "a should now be at the end of the list.");
    return gWalker.children(longlist);
  }).then(response => {
    is(nodeA, response.nodes[response.nodes.length - 1], "a should now be the last returned child.");
    // Now move it to the middle of the list.
    nextNode = response.nodes[13];
    return gWalker.insertBefore(nodeA, longlist, nextNode);
  }).then(response => {
    let sibling = inspector._documentWalker(gInspectee.querySelector("#a"), window).nextSibling();
    is(sibling, nextNode.rawNode(), "Node should match the expected next node.");
    return gWalker.children(longlist);
  }).then(response => {
    is(nodeA, response.nodes[13], "a should be where we expect it.");
    is(nextNode, response.nodes[14], "next node should be where we expect it.");
  }).then(runNextTest));
});

addTest(function cleanup() {
  delete gWalker;
  delete gClient;
  runNextTest();
});


  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
